Move Internal class into a distinct namespace
