+++ src/WhatWg/Url.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Uri\WhatWg;
 
+use Bakame\Aide\Helper\UrlValidationErrorCollector;
 use Exception;
 use League\Uri\Idna\Converter;
 use League\Uri\UriString;
